Add files via upload
Hand Written Digit Recognision System is a computer-based program that can recognize and categorize handwritten digits into their numerical values (0 through 9), an application that employs machine learning and its associated algorithms, particularly in the field of computer vision. For this project we have used MNIST dataset, a commonly used standard dataset in machine learning and computer vision. This dataset has been created in order to train and evaluate machine learning models for handwritten digit recognition. This dataset is made up of grayscale photographs of handwritten numerals 0 through 9. Each image has a height of 28 pixels and a width of 28 pixels, for a total of 784 pixels per image. Therefore, it a tiny dataset in comparison to more recent datasets, yet it remains an accepted reference.

In this project, I will be developing a Convolutional Neural Network (CNN), which will be a fundamental element of our Handwritten Digit Recognition project.  Along the way, I may run across a number of obstacles, such overfitting, disappearing gradients, or problems with training convergence. Debugging and solving these issues will be my responsibility. Overall, I aim to create an efficient CNN that can recognize a handwritten digit with maximum accuracy.